Overflow eliminated

+++ homeworks/HW1/src/main/java/ru.atom/Util.java
@@ -3,10 +3,9 @@ package ru.atom;
 /**
  * In this assignment you need to implement the following util methods.
  * Note:
- *  throw new UnsupportedOperationException(); - is just a stub
+ * throw new UnsupportedOperationException(); - is just a stub
  */
 public class Util {
-
 
 
     /**
@@ -17,7 +16,7 @@ public class Util {
      */
     public static int max(int[] values) {
         int maxVal = values[0];
-        for (int i = 0;i < values.length;i++) {
+        for (int i = 0; i < values.length; i++) {
             if (values[i] > maxVal) {
                 maxVal = values[i];
             }
@@ -32,7 +31,7 @@ public class Util {
      * @return the sum of all values.
      */
     public static long sum(int[] values) {
-        int sum = 0;
+        long sum = 0;
         for (int i = 0; i < values.length; i++) {
             sum += values[i];
         }
